requirements/base.txt
django>=1.8.8,<1.9
djangorestframework>=3.3,<3.4
markdown>=2.6,<2.7
django-filter>=0.12,<0.13
django-extensions>=1.6,<1.7
django-imagekit>=3.3,<3.4
Pillow>=3.2,<3.3
djangorestframework-registration>=0.0,<0.1
